requirements.txt
pydantic
python_dotenv
requests
aiohttp
aiofiles
httpx